+++ tourism_crawler/models/__init__.py
@@ -1,13 +1,5 @@
-from tourism_crawler.models.base import Base
-from tourism_crawler.models.place import PlaceModel
-from tourism_crawler.models.restaurant import RestaurantModel
-from tourism_crawler.models.coffee import CoffeeShopModel
-from tourism_crawler.models.homestay import HomestayModel
-from tourism_crawler.models.media import MediaItemModel
-from tourism_crawler.models.verification import VerificationSourceModel
+from tourism_crawler.models.place import PlaceRawModel, PlaceEnrichedModel
 
 __all__ = [
-    "Base", "PlaceModel", "RestaurantModel",
-    "CoffeeShopModel", "HomestayModel", "MediaItemModel",
-    "VerificationSourceModel"
+    "PlaceRawModel", "PlaceEnrichedModel"
 ]